biblioteca-demo-files/backend/requirements.txt
django==4.2
djangorestframework==3.14.0
django-cors-headers==4.3.0
